@@ -1,24 +1,25 @@
 package io.spring.api.security;
 
-import io.spring.core.service.JwtService;
-import io.spring.core.user.UserRepository;
+import io.spring.core.user.domain.UserRepository;
+import io.spring.infrastructure.service.JwtService;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.Optional;
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import org.springframework.beans.factory.annotation.Autowired;
+import java.util.UUID;
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.web.filter.OncePerRequestFilter;
 
-@SuppressWarnings("SpringJavaAutowiringInspection")
+@RequiredArgsConstructor
 public class JwtTokenFilter extends OncePerRequestFilter {
-  @Autowired private UserRepository userRepository;
-  @Autowired private JwtService jwtService;
+  private final UserRepository userRepository;
+  private final JwtService jwtService;
   private final String header = "Authorization";
 
   @Override
@@ -31,7 +32,7 @@ public class JwtTokenFilter extends OncePerRequestFilter {
             id -> {
               if (SecurityContextHolder.getContext().getAuthentication() == null) {
                 userRepository
-                    .findById(id)
+                    .findById(UUID.fromString(id))
                     .ifPresent(
                         user -> {
                           UsernamePasswordAuthenticationToken authenticationToken =
