@@ -1,14 +1,14 @@
-package io.spring.infrastructure.mybatis.mapper;
-
-import io.spring.core.comment.Comment;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
-
-@Mapper
-public interface CommentMapper {
-  void insert(@Param("comment") Comment comment);
-
-  Comment findById(@Param("articleId") String articleId, @Param("id") String id);
-
-  void delete(@Param("id") String id);
-}
+// package io.spring.infrastructure.mybatis.mapper;
+//
+// import io.spring.core.comment.domain.Comment;
+// import org.apache.ibatis.annotations.Mapper;
+// import org.apache.ibatis.annotations.Param;
+//
+// @Mapper
+// public interface CommentMapper {
+//   void insert(@Param("comment") Comment comment);
+//
+//   Comment findById(@Param("articleId") String articleId, @Param("id") String id);
+//
+//   void delete(@Param("id") String id);
+// }
