@@ -1,9 +1,9 @@
-package io.spring.infrastructure.mybatis.readservice;
-
-import java.util.List;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface TagReadService {
-  List<String> all();
-}
+// package io.spring.infrastructure.mybatis.readservice;
+//
+// import java.util.List;
+// import org.apache.ibatis.annotations.Mapper;
+//
+// @Mapper
+// public interface TagReadService {
+//   List<String> all();
+// }
