@@ -1,13 +1,13 @@
-package io.spring.infrastructure.mybatis.readservice;
-
-import io.spring.application.data.UserData;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
-
-@Mapper
-public interface UserReadService {
-
-  UserData findByUsername(@Param("username") String username);
-
-  UserData findById(@Param("id") String id);
-}
+// package io.spring.infrastructure.mybatis.readservice;
+//
+// import io.spring.api.user.response.UserData;
+// import org.apache.ibatis.annotations.Mapper;
+// import org.apache.ibatis.annotations.Param;
+//
+// @Mapper
+// public interface UserReadService {
+//
+//   UserData findByUsername(@Param("username") String username);
+//
+//   UserData findById(@Param("id") String id);
+// }
