@@ -1,11 +1,8 @@
 package io.spring;
 
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
+// @SpringBootTest
 public class RealworldApplicationTests {
 
-  @Test
+  //  @Test
   public void contextLoads() {}
 }
