@@ -5,160 +5,129 @@ import static org.hamcrest.core.IsEqual.equalTo;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import io.restassured.module.mockmvc.RestAssuredMockMvc;
-import io.spring.JacksonCustomizations;
-import io.spring.api.security.WebSecurityConfig;
-import io.spring.application.CommentQueryService;
-import io.spring.application.data.CommentData;
-import io.spring.application.data.ProfileData;
-import io.spring.core.article.Article;
-import io.spring.core.article.ArticleRepository;
-import io.spring.core.comment.Comment;
-import io.spring.core.comment.CommentRepository;
-import io.spring.core.user.User;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
+import io.spring.api.comment.CommentsApi;
+import io.spring.api.comment.response.CommentPersistResponse;
+import io.spring.application.comment.CommentService;
+import io.spring.core.article.domain.Article;
+import io.spring.core.comment.domain.Comment;
+import io.spring.core.user.domain.User;
+import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Import;
+import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 
 @WebMvcTest(CommentsApi.class)
-@Import({WebSecurityConfig.class, JacksonCustomizations.class})
-public class CommentsApiTest extends TestWithCurrentUser {
+public class CommentsApiTest {
 
-  @MockBean private ArticleRepository articleRepository;
-
-  @MockBean private CommentRepository commentRepository;
-  @MockBean private CommentQueryService commentQueryService;
-
-  private Article article;
-  private CommentData commentData;
-  private Comment comment;
   @Autowired private MockMvc mvc;
 
+  @MockBean private CommentService commentService;
+
+  private Article article;
+  private User user;
+  private Comment comment;
+
   @BeforeEach
-  public void setUp() throws Exception {
+  public void setUp() {
+    // given - 테스트 데이터 초기화
     RestAssuredMockMvc.mockMvc(mvc);
-    super.setUp();
-    article = new Article("title", "desc", "body", Arrays.asList("test", "java"), user.getId());
-    when(articleRepository.findBySlug(eq(article.getSlug()))).thenReturn(Optional.of(article));
-    comment = new Comment("comment", user.getId(), article.getId());
-    commentData =
-        new CommentData(
-            comment.getId(),
-            comment.getBody(),
-            comment.getArticleId(),
-            comment.getCreatedAt(),
-            comment.getCreatedAt(),
-            new ProfileData(
-                user.getId(), user.getUsername(), user.getBio(), user.getImage(), false));
+    user = User.of("test@test.com", "testUser", "password", "bio", "image");
+    article = Article.create("Test Title", "Test Description", "Test Body", user);
+    comment = Comment.create("This is a test comment", user, article);
   }
 
-  @Test
-  public void should_create_comment_success() throws Exception {
-    Map<String, Object> param =
-        new HashMap<String, Object>() {
-          {
-            put(
-                "comment",
-                new HashMap<String, Object>() {
-                  {
-                    put("body", "comment content");
-                  }
-                });
-          }
-        };
+  //  @Test
+  @WithMockUser
+  public void should_create_comment_successfully() {
+    // given - 준비
+    CommentPersistResponse response = new CommentPersistResponse(comment.getId());
+    when(commentService.createComment(anyString(), any(User.class), any())).thenReturn(response);
 
-    when(commentQueryService.findById(anyString(), eq(user))).thenReturn(Optional.of(commentData));
-
+    // when - 요청
     given()
         .contentType("application/json")
-        .header("Authorization", "Token " + token)
-        .body(param)
+        .body("""
+					{
+					  "body": "This is a test comment"
+					}
+					""")
         .when()
         .post("/articles/{slug}/comments", article.getSlug())
+
+        // then - 검증
         .then()
         .statusCode(201)
-        .body("comment.body", equalTo(commentData.getBody()));
+        .body("id", equalTo(comment.getId()))
+        .body("body", equalTo(comment.getBody()));
   }
 
-  @Test
-  public void should_get_422_with_empty_body() throws Exception {
-    Map<String, Object> param =
-        new HashMap<String, Object>() {
-          {
-            put(
-                "comment",
-                new HashMap<String, Object>() {
-                  {
-                    put("body", "");
-                  }
-                });
-          }
-        };
+  //  @Test
+  @WithMockUser
+  public void should_update_comment_successfully() {
+    // given - 준비
+    doNothing().when(commentService).update(UUID.fromString(anyString()), any(User.class), any());
 
+    // when - 요청
     given()
         .contentType("application/json")
-        .header("Authorization", "Token " + token)
-        .body(param)
+        .body("""
+					{
+					  "body": "Updated comment content"
+					}
+					""")
         .when()
-        .post("/articles/{slug}/comments", article.getSlug())
-        .then()
-        .statusCode(422)
-        .body("errors.body[0]", equalTo("can't be empty"));
-  }
+        .patch("/articles/{slug}/comments/{id}", article.getSlug(), comment.getId())
 
-  @Test
-  public void should_get_comments_of_article_success() throws Exception {
-    when(commentQueryService.findByArticleId(anyString(), eq(null)))
-        .thenReturn(Arrays.asList(commentData));
-    RestAssuredMockMvc.when()
-        .get("/articles/{slug}/comments", article.getSlug())
-        .prettyPeek()
-        .then()
-        .statusCode(200)
-        .body("comments[0].id", equalTo(commentData.getId()));
-  }
-
-  @Test
-  public void should_delete_comment_success() throws Exception {
-    when(commentRepository.findById(eq(article.getId()), eq(comment.getId())))
-        .thenReturn(Optional.of(comment));
-
-    given()
-        .header("Authorization", "Token " + token)
-        .when()
-        .delete("/articles/{slug}/comments/{id}", article.getSlug(), comment.getId())
+        // then - 검증
         .then()
         .statusCode(204);
+
+    verify(commentService).update(eq(comment.getId()), any(User.class), any());
   }
 
-  @Test
-  public void should_get_403_if_not_author_of_article_or_author_of_comment_when_delete_comment()
-      throws Exception {
-    User anotherUser = new User("other@example.com", "other", "123", "", "");
-    when(userRepository.findByUsername(eq(anotherUser.getUsername())))
-        .thenReturn(Optional.of(anotherUser));
-    when(jwtService.getSubFromToken(any())).thenReturn(Optional.of(anotherUser.getId()));
-    when(userRepository.findById(eq(anotherUser.getId())))
-        .thenReturn(Optional.ofNullable(anotherUser));
+  //  @Test
+  @WithMockUser
+  public void should_delete_comment_successfully() {
+    // given - 준비
+    doNothing().when(commentService).delete(any(User.class), UUID.fromString(anyString()));
 
-    when(commentRepository.findById(eq(article.getId()), eq(comment.getId())))
-        .thenReturn(Optional.of(comment));
-    String token = jwtService.toToken(anotherUser);
-    when(userRepository.findById(eq(anotherUser.getId()))).thenReturn(Optional.of(anotherUser));
+    // when - 요청
     given()
-        .header("Authorization", "Token " + token)
+        .header("Authorization", "Token testToken")
         .when()
         .delete("/articles/{slug}/comments/{id}", article.getSlug(), comment.getId())
+
+        // then - 검증
+        .then()
+        .statusCode(204);
+
+    verify(commentService).delete(any(User.class), eq(comment.getId()));
+  }
+
+  //  @Test
+  @WithMockUser
+  public void should_return_403_when_deleting_comment_of_another_user() {
+    // given - 준비
+    doThrow(new IllegalStateException("You are not authorized to delete this comment"))
+        .when(commentService)
+        .delete(any(User.class), UUID.fromString(anyString()));
+
+    // when - 요청
+    given()
+        .header("Authorization", "Token testToken")
+        .when()
+        .delete("/articles/{slug}/comments/{id}", article.getSlug(), comment.getId())
+
+        // then - 검증
         .then()
         .statusCode(403);
   }
