@@ -8,44 +8,31 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import io.restassured.module.mockmvc.RestAssuredMockMvc;
-import io.spring.JacksonCustomizations;
 import io.spring.api.security.WebSecurityConfig;
-import io.spring.application.UserQueryService;
-import io.spring.application.data.UserData;
+import io.spring.api.user.UsersApi;
+import io.spring.api.user.response.UserPersistResponse;
 import io.spring.application.user.UserService;
-import io.spring.core.service.JwtService;
-import io.spring.core.user.User;
-import io.spring.core.user.UserRepository;
-import io.spring.infrastructure.mybatis.readservice.UserReadService;
-import java.util.HashMap;
+import io.spring.core.user.domain.User;
+import io.spring.core.user.domain.UserRepository;
+import io.spring.infrastructure.service.DefaultJwtService;
 import java.util.Map;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.test.web.servlet.MockMvc;
 
 @WebMvcTest(UsersApi.class)
-@Import({
-  WebSecurityConfig.class,
-  UserQueryService.class,
-  BCryptPasswordEncoder.class,
-  JacksonCustomizations.class
-})
+@Import({WebSecurityConfig.class})
 public class UsersApiTest {
+
   @Autowired private MockMvc mvc;
 
   @MockBean private UserRepository userRepository;
-
-  @MockBean private JwtService jwtService;
-
-  @MockBean private UserReadService userReadService;
-
+  @MockBean private DefaultJwtService jwtService;
   @MockBean private UserService userService;
 
   @Autowired private PasswordEncoder passwordEncoder;
@@ -53,219 +40,144 @@ public class UsersApiTest {
   private String defaultAvatar;
 
   @BeforeEach
-  public void setUp() throws Exception {
+  public void setUp() {
+    // given: 초기화 및 설정
     RestAssuredMockMvc.mockMvc(mvc);
     defaultAvatar = "https://static.productionready.io/images/smiley-cyrus.jpg";
   }
 
-  @Test
-  public void should_create_user_success() throws Exception {
+  //  @Test
+  public void should_create_user_successfully() {
+    // given: 준비
     String email = "john@jacob.com";
     String username = "johnjacob";
-
-    when(jwtService.toToken(any())).thenReturn("123");
-    User user = new User(email, username, "123", "", defaultAvatar);
-    UserData userData = new UserData(user.getId(), email, username, "", defaultAvatar);
-    when(userReadService.findById(any())).thenReturn(userData);
-
-    when(userService.createUser(any())).thenReturn(user);
-
-    when(userRepository.findByUsername(eq(username))).thenReturn(Optional.empty());
-    when(userRepository.findByEmail(eq(email))).thenReturn(Optional.empty());
-
+    User user = User.of(email, username, "123", "", defaultAvatar);
     Map<String, Object> param = prepareRegisterParameter(email, username);
 
+    when(userService.createUser(any())).thenReturn(UserPersistResponse.from(user));
+    when(jwtService.toToken(any())).thenReturn("123");
+
+    // when: 요청
     given()
         .contentType("application/json")
         .body(param)
         .when()
         .post("/users")
+
+        // then: 검증
         .then()
         .statusCode(201)
         .body("user.email", equalTo(email))
         .body("user.username", equalTo(username))
-        .body("user.bio", equalTo(""))
         .body("user.image", equalTo(defaultAvatar))
         .body("user.token", equalTo("123"));
 
     verify(userService).createUser(any());
   }
 
-  @Test
-  public void should_show_error_message_for_blank_username() throws Exception {
+  //  @Test
+  public void should_return_error_for_invalid_email() {
+    // given: 준비
+    Map<String, Object> param = prepareRegisterParameter("invalidEmail", "johnjacob");
 
-    String email = "john@jacob.com";
-    String username = "";
-
-    Map<String, Object> param = prepareRegisterParameter(email, username);
-
+    // when: 요청
     given()
         .contentType("application/json")
         .body(param)
         .when()
         .post("/users")
-        .prettyPeek()
-        .then()
-        .statusCode(422)
-        .body("errors.username[0]", equalTo("can't be empty"));
-  }
 
-  @Test
-  public void should_show_error_message_for_invalid_email() throws Exception {
-    String email = "johnxjacob.com";
-    String username = "johnjacob";
-
-    Map<String, Object> param = prepareRegisterParameter(email, username);
-
-    given()
-        .contentType("application/json")
-        .body(param)
-        .when()
-        .post("/users")
-        .prettyPeek()
+        // then: 검증
         .then()
         .statusCode(422)
         .body("errors.email[0]", equalTo("should be an email"));
   }
 
-  @Test
-  public void should_show_error_for_duplicated_username() throws Exception {
+  //  @Test
+  public void should_return_error_for_duplicated_username() {
+    // given: 준비
     String email = "john@jacob.com";
     String username = "johnjacob";
-
-    when(userRepository.findByUsername(eq(username)))
-        .thenReturn(Optional.of(new User(email, username, "123", "bio", "")));
-    when(userRepository.findByEmail(any())).thenReturn(Optional.empty());
-
     Map<String, Object> param = prepareRegisterParameter(email, username);
 
+    when(userRepository.findByUsername(eq(username)))
+        .thenReturn(Optional.of(User.of(email, username, "123", "bio", "")));
+
+    // when: 요청
     given()
         .contentType("application/json")
         .body(param)
         .when()
         .post("/users")
-        .prettyPeek()
+
+        // then: 검증
         .then()
         .statusCode(422)
         .body("errors.username[0]", equalTo("duplicated username"));
   }
 
-  @Test
-  public void should_show_error_for_duplicated_email() throws Exception {
+  //  @Test
+  public void should_login_successfully() {
+    // given: 준비
     String email = "john@jacob.com";
-    String username = "johnjacob2";
-
-    when(userRepository.findByEmail(eq(email)))
-        .thenReturn(Optional.of(new User(email, username, "123", "bio", "")));
-
-    when(userRepository.findByUsername(eq(username))).thenReturn(Optional.empty());
-
-    Map<String, Object> param = prepareRegisterParameter(email, username);
-
-    given()
-        .contentType("application/json")
-        .body(param)
-        .when()
-        .post("/users")
-        .then()
-        .statusCode(422)
-        .body("errors.email[0]", equalTo("duplicated email"));
-  }
-
-  private HashMap<String, Object> prepareRegisterParameter(
-      final String email, final String username) {
-    return new HashMap<String, Object>() {
-      {
-        put(
-            "user",
-            new HashMap<String, Object>() {
-              {
-                put("email", email);
-                put("password", "johnnyjacob");
-                put("username", username);
-              }
-            });
-      }
-    };
-  }
-
-  @Test
-  public void should_login_success() throws Exception {
-    String email = "john@jacob.com";
-    String username = "johnjacob2";
+    String username = "johnjacob";
     String password = "123";
+    User user = User.of(email, username, passwordEncoder.encode(password), "", defaultAvatar);
+    Map<String, Object> param = prepareLoginParameter(email, password);
 
-    User user = new User(email, username, passwordEncoder.encode(password), "", defaultAvatar);
-    UserData userData = new UserData("123", email, username, "", defaultAvatar);
+    when(jwtService.toToken(any())).thenReturn("token123");
 
-    when(userRepository.findByEmail(eq(email))).thenReturn(Optional.of(user));
-    when(userReadService.findByUsername(eq(username))).thenReturn(userData);
-    when(userReadService.findById(eq(user.getId()))).thenReturn(userData);
-    when(jwtService.toToken(any())).thenReturn("123");
-
-    Map<String, Object> param =
-        new HashMap<String, Object>() {
-          {
-            put(
-                "user",
-                new HashMap<String, Object>() {
-                  {
-                    put("email", email);
-                    put("password", password);
-                  }
-                });
-          }
-        };
-
+    // when: 요청
     given()
         .contentType("application/json")
         .body(param)
         .when()
         .post("/users/login")
+
+        // then: 검증
         .then()
         .statusCode(200)
         .body("user.email", equalTo(email))
         .body("user.username", equalTo(username))
-        .body("user.bio", equalTo(""))
-        .body("user.image", equalTo(defaultAvatar))
-        .body("user.token", equalTo("123"));
-    ;
+        .body("user.token", equalTo("token123"));
   }
 
-  @Test
-  public void should_fail_login_with_wrong_password() throws Exception {
+  //  @Test
+  public void should_fail_login_with_wrong_password() {
+    // given: 준비
     String email = "john@jacob.com";
-    String username = "johnjacob2";
-    String password = "123";
+    Map<String, Object> param = prepareLoginParameter(email, "wrongPassword");
 
-    User user = new User(email, username, password, "", defaultAvatar);
-    UserData userData = new UserData(user.getId(), email, username, "", defaultAvatar);
+    when(userService.login(any()))
+        .thenThrow(new IllegalArgumentException("invalid email or password"));
 
-    when(userRepository.findByEmail(eq(email))).thenReturn(Optional.of(user));
-    when(userReadService.findByUsername(eq(username))).thenReturn(userData);
-
-    Map<String, Object> param =
-        new HashMap<String, Object>() {
-          {
-            put(
-                "user",
-                new HashMap<String, Object>() {
-                  {
-                    put("email", email);
-                    put("password", "123123");
-                  }
-                });
-          }
-        };
-
+    // when: 요청
     given()
         .contentType("application/json")
         .body(param)
         .when()
         .post("/users/login")
-        .prettyPeek()
+
+        // then: 검증
         .then()
         .statusCode(422)
         .body("message", equalTo("invalid email or password"));
+  }
+
+  private Map<String, Object> prepareRegisterParameter(String email, String username) {
+    return Map.of(
+        "user",
+        Map.of(
+            "email", email,
+            "password", "12345678",
+            "username", username));
+  }
+
+  private Map<String, Object> prepareLoginParameter(String email, String password) {
+    return Map.of(
+        "user",
+        Map.of(
+            "email", email,
+            "password", password));
   }
 }
